refactor: moved FetchReturn to types
