+++ src/app/dashboard/users/MembersTable.tsx
@@ -1,9 +1,8 @@
 'use client';
 
+import { FetchReturn, OrganizationMember } from '@/lib/types';
 import { ColumnDef } from '@tanstack/react-table';
 import DataTable from '@/components/shared/DataTable';
-import { FetchReturn } from '@/lib/customFetch';
-import { OrganizationMember } from '@/lib/types';
 import React from 'react';
 import { Skeleton } from '@/components/ui/skeleton';
