+++ src/components/providers/OrganizationContextProvider.tsx
@@ -1,8 +1,7 @@
 'use client';
 
+import { FetchReturn, OrganizationData } from '@/lib/types';
 import React, { createContext, useMemo } from 'react';
-import { FetchReturn } from '@/lib/customFetch';
-import { OrganizationData } from '@/lib/types';
 
 export interface IOrganizationContext extends Partial<OrganizationData> {
   _count?: { users: number; tickets: number };
